requirements.txt
fastapi[all]
starlette
pydantic
supabase
passlib
authlib
python-dotenv
tensorflow
numpy
google-generativeai
jinja2
uvicorn
pydantic[email]
bcrypt
pillow